[pvr] move jobs from PVRManager to PVRJobs class
