+++ xbmc/pvr/PVRJobs.cpp
@@ -23,8 +23,16 @@
 #include "dialogs/GUIDialogKaiToast.h"
 #include "events/EventLog.h"
 #include "events/NotificationEvent.h"
+#include "interfaces/AnnouncementManager.h"
 
+#include "PlayListPlayer.h"
 #include "pvr/PVRGUIActions.h"
+#include "pvr/PVRManager.h"
+#include "pvr/addons/PVRClients.h"
+#include "pvr/channels/PVRChannelGroupsContainer.h"
+#include "pvr/recordings/PVRRecordings.h"
+#include "pvr/timers/PVRTimers.h"
+#include "ServiceBroker.h"
 
 namespace PVR
 {
@@ -62,6 +70,72 @@ bool CPVREventlogJob::DoWork()
       EventPtr(new CNotificationEvent(event.m_label, event.m_msg, event.m_icon, event.m_bError ? EventLevel::Error : EventLevel::Information)));
   }
   return true;
+}
+
+bool CPVRChannelSwitchJob::DoWork(void)
+{
+  // announce OnStop and delete m_previous when done
+  if (m_previous)
+  {
+    CVariant data(CVariant::VariantTypeObject);
+    data["end"] = true;
+    ANNOUNCEMENT::CAnnouncementManager::GetInstance().Announce(ANNOUNCEMENT::Player, "xbmc", "OnStop", CFileItemPtr(m_previous), data);
+  }
+
+  // announce OnPlay if the switch was successful
+  if (m_next)
+  {
+    CVariant param;
+    param["player"]["speed"] = 1;
+    param["player"]["playerid"] = CServiceBroker::GetPlaylistPlayer().GetCurrentPlaylist();
+    ANNOUNCEMENT::CAnnouncementManager::GetInstance().Announce(ANNOUNCEMENT::Player, "xbmc", "OnPlay", CFileItemPtr(new CFileItem(*m_next)), param);
+  }
+
+  return true;
+}
+
+bool CPVRSearchMissingChannelIconsJob::DoWork(void)
+{
+  CPVRManager::GetInstance().SearchMissingChannelIcons();
+  return true;
+}
+
+bool CPVRClientConnectionJob::DoWork(void)
+{
+  CPVRManager::GetInstance().Clients()->ConnectionStateChange(m_client, m_connectString, m_state, m_message);
+  return true;
+}
+
+bool CPVRStartupJob::DoWork(void)
+{
+  CPVRManager::GetInstance().Clients()->Start();
+  return true;
+}
+
+bool CPVREpgsCreateJob::DoWork(void)
+{
+  return CPVRManager::GetInstance().CreateChannelEpgs();
+}
+
+bool CPVRRecordingsUpdateJob::DoWork(void)
+{
+  CPVRManager::GetInstance().Recordings()->Update();
+  return true;
+}
+
+bool CPVRTimersUpdateJob::DoWork(void)
+{
+  return CPVRManager::GetInstance().Timers()->Update();
+}
+
+bool CPVRChannelsUpdateJob::DoWork(void)
+{
+  return CPVRManager::GetInstance().ChannelGroups()->Update(true);
+}
+
+bool CPVRChannelGroupsUpdateJob::DoWork(void)
+{
+  return CPVRManager::GetInstance().ChannelGroups()->Update(false);
 }
 
 } // namespace PVR
